Update issues table row detection due to internal changes on DataTables 1.10.23.
